+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 IMPROVEMENTS
 
-* Migrate to the official mcp go-sdk
+* Migrate to the official mcp go-sdk [348](https://github.com/hashicorp/terraform-mcp-server/pull/348)
 
 # 1.0.0